moved all context objects into single object
