+++ src/components/providers/game/index.jsx
@@ -1,10 +1,11 @@
 import React, { useReducer } from 'react';
+import PropTypes from 'prop-types';
 import { GameContext } from 'contexts/game';
-import { baseState } from 'contexts/game';
+import { baseState } from 'contexts';
 import { gameReducer } from 'contexts/game/reducer';
 
 const GameProvider = ({ children, providedState = null } = {}) => {
-  const initialState = providedState || baseState;
+  const initialState = providedState || baseState.game;
   const [state, dispatch] = useReducer(gameReducer, initialState);
 
   return (
@@ -12,6 +13,11 @@ const GameProvider = ({ children, providedState = null } = {}) => {
       {children}
     </GameContext.Provider>
   );
+};
+
+GameProvider.propTypes = {
+  children: PropTypes.node,
+  providedState: PropTypes.object,
 };
 
 export default GameProvider;
